Add basic timing thread implementation
It isn't very dynamic yet, i.e., parameters can't be changed on-the-fly.
But it should be steadier than the old code already!

The one user-visible difference other than steady timing should be that
the "TAP" button no longer makes sound when clicked.  I made the
executive decision to remove that sound, since it would complicate the
implementation to keep it now, and most applications that I've seen are
silent when clicking their tap tempo buttons.